move: resume 파일 경로 이동
