--- backend/src/main/java/com/pofo/backend/domain/resume/repository/ResumeRepository.java
@@ -1,8 +1,0 @@
-package com.pofo.backend.domain.resume.repository;
-
-import com.pofo.backend.domain.resume.entity.Resume;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ResumeRepository extends JpaRepository<Resume, Integer> {
-
-}
