+++ backend/src/main/java/com/pofo/backend/domain/resume/resume/dto/request/ResumeCreateRequest.java
@@ -1,4 +1,4 @@
-package com.pofo.backend.domain.resume.dto.request;
+package com.pofo.backend.domain.resume.resume.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
