+++ backend/src/main/java/com/pofo/backend/domain/resume/resume/exception/ResumeCreationException.java
@@ -1,4 +1,4 @@
-package com.pofo.backend.domain.resume.exception;
+package com.pofo.backend.domain.resume.resume.exception;
 
 public class ResumeCreationException extends RuntimeException {
     public ResumeCreationException(String message) {
